requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pytest==8.4.1
httpx==0.28.1
openai==1.102.0
python-dotenv==1.1.1
python-multipart==0.0.20
